Import DSM 7.0-41890 armada375 linux-3.x.txz Base on Linux Kernal v3.2.101
tar url: https://global.synologydownload.com/download/ToolChain/Synology%20NAS%20GPL%20Source/7.0-41890/armada375/linux-3.x.txz
tar md5sum: 976ed5c8d8dfdfca74ddb3e81e6ba00a
from: https://archive.synology.com/download/ToolChain/Synology%20NAS%20GPL%20Source/7.0-41890

+++ include/linux/raid/libmd-report.h
@@ -1,14 +1,14 @@
 #ifndef MY_ABC_HERE
 #define MY_ABC_HERE
 #endif
- 
+// Copyright (c) 2000-2008 Synology Inc. All rights reserved.
 #ifndef _LIBMD_REPORT_H
 #define _LIBMD_REPORT_H
 
 #ifdef MY_ABC_HERE
 
 extern int (*funcSYNOSendRaidEvent)(unsigned int type, unsigned int raidno,
-									unsigned int diskno, unsigned int sector);
+									unsigned int diskno, unsigned long long sector);
 
 void SynoReportBadSector(sector_t sector, unsigned long rw, 
 						 int md_minor, struct block_device *bdev, const char *szFuncName);
@@ -17,11 +17,11 @@ void SynoReportCorrectBadSector(sector_t sector, int md_minor,
 								struct block_device *bdev, const char *szFuncName);
 
 void SynoReportFaultyDevice(int md_minor, struct block_device *bdev);
-#endif  
+#endif /* MY_ABC_HERE */
 
 #ifdef MY_ABC_HERE
 extern int (*funcSYNOSendAutoRemapLVMEvent)(const char*, unsigned long long, unsigned int);
 extern int (*funcSYNOSendAutoRemapRaidEvent)(unsigned int, unsigned long long, unsigned int);
 #endif
 
-#endif  
+#endif /* _LIBMD_REPORT_H */
